requirements.txt
# actionlib==1.14.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
# angles==1.9.13
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.13.0
# catkin==0.8.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
# diagnostic-updater==1.11.0
# dynamic-reconfigure==1.7.3
exceptiongroup==1.2.2
filelock==3.16.1
flask==3.0.3
Flask-Cors==5.0.0
flask-session==0.8.0
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
# gencpp==0.7.0
# geneus==3.0.0
# genlisp==0.4.18
# genmsg==0.6.0
# gennodejs==2.0.2
# genpy/==0.6.15
# geodesy==0.5.6
# gps-common==0.3.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
inexactsearch==1.0.2
# interactive-markers==1.12.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.2.17
langchain-core==0.2.43
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langsmith==0.1.147
# laser-geometry==1.6.7
MarkupSafe==2.1.5
matplotlib==3.7.5
# message-filters==1.17.0
mpmath==1.3.0
msgspec==0.18.6
multidict==6.1.0
networkx==3.1
# numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
orjson==3.10.12
packaging==24.2
pandas==2.0.3
pillow==10.4.0
polars==1.8.2
propcache==0.2.0
pyarrow==17.0.0
pydantic==2.10.3
pydantic-core==2.27.1
pyparsing==3.1.4
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
# python-qt-binding==0.4.4
pytz==2024.2
PyYAML==6.0.2
# qt-dotgraph==0.4.2
# qt-gui==0.4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
# resource-retriever==1.12.7
# rosbag==1.17.0
# rosclean==1.15.8/
# rosgraph==1.17.0
# roslaunch==1.17.0
# roslib==1.15.8
# roslint==0.12.0/
# roslz4==1.17.0/
# rosmaster==1.17.0
# rosmsg==1.17.0
# rosnode==1.17.0
# rosparam==1.17.0
# rospy==1.17.0
# rosservice==1.17.0
# rostest==1.17.0//
# rostopic==1.17.0
# rosunit==1.15.8
# roswtf==1.17.0
# rqt-graph==0.4.14
# rqt-gui==0.5.3
# rqt-gui-py==0.5.3
# rqt-tf-tree==0.6.4
# rviz==1.14.25
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
# sensor-msgs==1.13.1
sentence-transformers==3.2.1
silpa-common==0.3
six==1.16.0
# smclib==1.8.6
sniffio==1.3.1
soundex==1.1.3
soupsieve==2.6
spellchecker==0.4
SQLAlchemy==2.0.36
sympy==1.13.3
tenacity==8.5.0
# tf==1.13.2
# tf-conversio/ns/==1.13.2
# tf2-geometry-msgs==0.7.7
# tf2-py==0.7.7
# tf2-ros==0.7.7
threadpoolctl==3.5.0
tokenizers==0.20.3
# topic-tools==1.17.0
# torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
# triton==3.0.0
# turtlebot3-teleop==1.2.5
# typing-extensions==4.12.2
# tzdata==2024.2
# unique-id==1.0.6
urllib3==2.2.3
werkzeug==3.0.6
xgboost==2.1.3
yarl==1.15.2
zipp==3.20.2
